+++ lib/task/generator/skeleton/module/module/actions/actions.class.php
@@ -10,11 +10,11 @@
  */
 class ##MODULE_NAME##Actions extends sfActions
 {
- /**
-  * Executes index action
-  *
-  * @param sfRequest $request A request object
-  */
+  /**
+   * Executes index action
+   *
+   * @param sfRequest $request A request object
+   */
   public function executeIndex(sfWebRequest $request)
   {
     $this->forward('default', 'module');