+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 CHANGELOG
 =========
 
+ * Added skip-build option to sfDoctrineCreateModelTablesTask (patch from @estahn)
  * better code coverage for sfWebRequest class unit tests
  * added parameters proxy to sfWebRequest::getClientIp() method (default true)
  * introduced trust_proxy option on sfWebRequest (default true)